♻️ Refactor toggleOptionButton
