+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import options from '../assets/options.json'
 import { goto } from '../lib/router'
-import { renderOptions } from '../components/elements/option'
+import { renderOptions, toggleOptionButton } from '../components/elements/option'
 
 function renderIndex() {
   document.querySelector('#app').innerHTML = `
@@ -13,32 +13,17 @@ function renderIndex() {
   const optionsElement = document.querySelector('#options')
   renderOptions(optionsElement)
 
-  const selectedOptions = options.reduce((acc, option) => {
-    acc[option.id] = []
-    return acc
-  }, {})
   const submitButton = document.querySelector('#submit-button')
   submitButton.addEventListener('click', () => {
     goto(`/result?query=${JSON.stringify(selectedOptions)}`)
   })
 
-  const selectionButton = document.querySelectorAll('.selection-button')
-  Array.from(selectionButton).forEach((button) => {
-    button.addEventListener('click', () => {
-      const option = button.getAttribute('data-option-id')
-      const selection = button.getAttribute('data-selection-id')
-
-      if (selectedOptions[option].includes(selection)) {
-        selectedOptions[option] = selectedOptions[option].filter((item) => item !== selection)
-        button.classList.remove('blue-button')
-        button.classList.add('secondary-button')
-      } else {
-        selectedOptions[option].push(selection)
-        button.classList.remove('secondary-button')
-        button.classList.add('blue-button')
-      }
-    })
-  })
+  const selectionButtons = document.querySelectorAll('.selection-button')
+  let selectedOptions = options.reduce((acc, option) => {
+    acc[option.id] = []
+    return acc
+  }, {})
+  toggleOptionButton(selectionButtons, selectedOptions)
 }
 
 export { renderIndex }
